SquishAppFormat/pyproject.toml
[tool.poetry]
name = "SquishAppFormat"
version = "0.1.0"
description = ""
authors = ["Henry Post <henryfbp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
GitPython = "^3.1.31"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
